+++ plugin/src/main/scala/io.github.jonas.paradox.material.theme/ParadoxMaterialTheme.scala
@@ -10,17 +10,29 @@ case class ParadoxMaterialTheme(properties: Map[String, String]) {
   def withLanguage(locale: Locale) =
     withProperties("language" -> locale.getLanguage)
 
-  def withColor(primary: String, accent: String) =
-    withProperties(
+  def withColor(primaryColor: String, accentColor: String) = {
+    val primary = primaryColor.replace(" ", "-").toLowerCase
+    val accent = accentColor.replace(" ", "-").toLowerCase
+    val colorProps = withProperties(
       "color.primary" -> primary,
-      "color.accent"  -> accent
+      "color.accent" -> accent
     )
+    ParadoxMaterialTheme.Palette.Primary.get(primary) match {
+      case Some(themeColor) => colorProps.withProperties("color.primary.theme" -> themeColor)
+      case None             => colorProps
+    }
+  }
 
   def withFavicon(favicon: String) =
     withProperties("favicon" -> favicon)
 
-  def withLogo(logo: String) =
-    withProperties("logo" -> logo).withoutProperties("logo.icon")
+  def withLogo(logo: String) = {
+    val logoProps = withProperties("logo" -> logo).withoutProperties("logo.icon")
+    if (logo.startsWith("http"))
+      logoProps.withProperties("logo.http" -> logo)
+    else
+      logoProps
+  }
 
   def withLogoIcon(icon: String) =
     withProperties("logo.icon" -> icon).withoutProperties("logo")
@@ -101,5 +113,47 @@ object ParadoxMaterialTheme {
       .withLogoIcon("local_library")
       .withFavicon("assets/images/favicon.png")
       .withSearch()
+  }
+
+  object Palette {
+    val Primary = Map(
+      "red"         -> "#ef5350",
+      "pink"        -> "#e91e63",
+      "purple"      -> "#ab47bc",
+      "deep-purple" -> "#7e57c2",
+      "indigo"      -> "#3f51b5",
+      "blue"        -> "#2196f3",
+      "light-blue"  -> "#03a9f4",
+      "cyan"        -> "#00bcd4",
+      "teal"        -> "#009688",
+      "green"       -> "#4caf50",
+      "light-green" -> "#7cb342",
+      "lime"        -> "#c0ca33",
+      "yellow"      -> "#f9a825",
+      "amber"       -> "#ffa000",
+      "orange"      -> "#fb8c00",
+      "deep-orange" -> "#ff7043",
+      "brown"       -> "#795548",
+      "grey"        -> "#757575",
+      "blue-grey"   -> "#546e7a"
+    )
+    val Accent = Map(
+      "red"         -> "#ff1744",
+      "pink"        -> "#f50057",
+      "purple"      -> "#e040fb",
+      "deep-purple" -> "#7c4dff",
+      "indigo"      -> "#536dfe",
+      "blue"        -> "#448aff",
+      "light-blue"  -> "#0091ea",
+      "cyan"        -> "#00b8d4",
+      "teal"        -> "#00bfa5",
+      "green"       -> "#00c853",
+      "light-green" -> "#64dd17",
+      "lime"        -> "#aeea00",
+      "yellow"      -> "#ffd600",
+      "amber"       -> "#ffab00",
+      "orange"      -> "#ff9100",
+      "deep-orange" -> "#ff6e40"
+    )
   }
 }